Refactor overview to support `react-router` v6
